[ADD] Se agregan artefactos de componentes y validación de líneas en llamado a API

+++ dockerfiles/fastapi/app/schemas/linea.py
@@ -1,15 +1,16 @@
-from pydantic import BaseModel
 from datetime import date
 from typing import List
+
+from pydantic import BaseModel
+
 
 class LineaInput(BaseModel):
     linea: str
     date_init: date # Formato ISO 8601, por ejemplo: "2025-08-22"
 
-
 class FechaValor(BaseModel):
     fecha: date # Formato ISO 8601, por ejemplo: "2025-08-22"
-    valor: float
+    valor: int
 
 class LineaOutput(BaseModel):
     resultados: List[FechaValor]
